REMOVED `friendship_search_count` concept
Due to the fact that now we simply can retrieve `follower_count` and
`following_count` , we dont need the concept of an overall limit of
follower / following retrieving maximum. The necessary limits are now implemented
with `scan_limit` to control how many followers to scan and `reverse_following_count`
to control the maximum of followings to trace back when executing the `CategorizeAccountsStage`.