[fix] Fix off-by-one error in `FileAdapter.get_lines`
`FileAdapter.get_lines` get an optional parameter called `limit`. When specified,
it has returned one more than specified by `limit` due to the check:
`if(line_counter>limit)`. It dont check if they are equal which also means that the
limit is reached.
This commit fixes this by adding an eqauality check:
`if(line_counter>=limit)`.